requirements.txt
numpy
pandas
openai
pillow
requests
python-dotenv
pyzbar
zxing
openfoodfacts
fastapi
uvicorn
opencv-python